requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.7
django-cache-url==3.4.4
django-rest-framework==0.1.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.20.1
packaging==23.2
python-dotenv==1.0.0
pytz==2023.3.post1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
